moved grid to its own file
